Re-enable metadata collectors
